Have trezor ignore inputs it cannot sign rather than throwing an error
Instead of erroring and failing to sign the transaction when an input
cannot be signed, just ignore that input. Ideally this would be done
using InputScriptType.EXTERNAL, but the firmware does not support this
So to achieve this, we instead fill in the input with bogus data and
let the Trezor sign it anyways. Then we just ignore the signature
that was produced.