Introduce and implement prompt_pin and send_pin methods
Only the Trezor and KeepKey need to have a PIN entered. However this
PIN entering is canceled if the device is re-initialized which happens
every time the client is created. To work around this, a client
subclass for each is introduced which does not initialize the device
on creation. Instead device initialization is done at the beginning
of each call which requires it. At that time, a check is also done
to ensure that the device has the PIN cached. This allows for
send_pin to actually be able to send the PIN after promptpin.