Replace distutils.core with setuptools in setup.py
Poetry geneates a setup.py with distutils.core for the sdist
package, which is fine for installing with pip as pip will use
setuptools instead of distutils.core. But when people install
from source using the setup.py, distutils.core will not install
the dependencies; we need setuptools for that. So in our generated
setup.py file, replace distutils.core with setuptools.