Add get fingerprint function to HardwareWalletClient class
As pre-work for the bitbox02 integration, add a function to the device
class to get the master fingerprint. It can later be overloaded by the
bitbox02 definitions.

The existing hardware wallet clients now use this function to get the
fingerprint. Coldcard's existing fingerprint function is renamed to
overload the function definition.

Also removes the `get_xpub_fingerprint_as_id` function, since it does
the same as `get_xpub_fingerprint_hex`.