tests: Use descriptor wallets
We are really targeting descriptor wallets, so we should be testing
those.

Because descriptor wallets also change how the imports will work w.r.t
rescans and the next address to give out, the getkeypool path checks are
changed to just check that the begin with the correct path prefix.