bitbox02: implement multisig display_address()
If the multisig setup is not yet registered on the device, the user is
prompted to register and verify the setup on the device.

The descriptor in display_address is used to create a KeyOriginInfo
mapping. This format is chosen as it is also used in a PSBT object,
which will make the signing code easier.

Only p2wsh (native segwit) and p2wsh-p2sh (p2sh wrapped segwit
multisig) is supported. Legacy `p2wsh` multisig is not supported.