Change display_*_address to return address string
Instead of returning a dictionary for
HardwareWalletClient.display_single_sig_address and
HardwareWalletClient.display_multisig_address, return an address string
instead. The conversion to a dictionary for cli output is done by the
command handler in commands.py.

Also updates the functions to type check and the docstring in the
abstract class.