ledger: require the app name is Bitcoin or Bitcoin Test
Checks the app name and makes sure it is only Bitcoin or Bitcoin Test.
For older firmware that do not implement this command, we receive None.
In that case, we skip the check rather than disallow the device.

For Speculos, the app name is "app", so that is included as well.