deps: Remove python <3.10 restriction
This restriction was because pyside2 has a restriction on python <3.10.
As pyside2 is optional, and the binary builds are done with 3.7, it is
okay to remove this restriction globally and just limit it to pyside2
itself. So pyside2 will only be installed if the qt extra is specified,
and python <3.10.