keepkey, trezor: Handle if ButtonRequest is sent during sendpin
Keepkey is now displaying the passphrase to users after it receives it
and responding with a ButtonRequest message. However this was not being
properly handled by sendpin which would cause the device to hang as it
waited for a ButtonAck. To resolve this, use self.call but avoid the
firmware check which will fail due to the lack of self.features.