ledger: Try legacy protocol when unable to sign multisigs
When we have a multisig that cannot be signed with the new protocol, we
want to try with the old protocol as the old protocol was more lenient.

However, this makes the testing matrix stupidly complicated as the old
protocol does not support Taproot, so this behavior IS NOT TESTED. If a
Taproot output is present in the transaction, the device will throw an
error.