jade: Pass utxo satoshi value to Jade on signing
Pass the utxo value, along with the prior tx (non-witness-utxo) when it
is available.  Most of the time Jade requires the prior tx but there are
cases when it can be omitted if the sats value is passed.
Pass both whenever possible - the precise logic of which is required can
remain contained in the Jade fw.
(Improves compatibility with PSBT which does not always contain
'non-witness-utxo' for segwit inputs.)