jade: fix CI cached python env issue
Reinstall the esp-idf tools and re-create the python virtualenv for
esp-idf builds on every run, not just the runs that re-clone the esp-idf
repo.  Skipping this step appears to cause builds which get everything
from the work cache to not use the virtualenv, but to use the system
python, which then fails (as dependencies are not installed).