fix base58.decode_check()
The function used the full sha256 and not just the first four bytes to
verify the checksum. As a result, the checksum check always fails.

This breaks the BitBox02 multisig registration, which uses this
function to decode xpubs. This broke when it switched to using the
internal base58 library over the `base58` dependency in
674cb08d6ffcecbf2502a05dd4371c1cfc8baa93.

Fixes #700.