add support for Python 3.12
Fedora 39 ships with Python 3.12. To achieve a better
out-of-the-box experience, support for Python 3.12 is
added. This required `pyinstaller` and `hidapi` to
be updated.

`pyinstaller` has dropped support for Python 3.7 because
it reached EOL on on 2023-06-27 [1]. The `pyinstaller`
update forces us to also drop support for Python 3.7.
Considering that HWI is a security-sensitive package it
makes sense to stop using EOL software components.

[1] https://github.com/pyinstaller/pyinstaller/commit/da33dbf1681afa7541124034a8aa40227353dcf2